Create own module.
