+++ src/estimagic/optimization/tranquilo/geometry.py
@@ -1,7 +1,6 @@
 from functools import partial
 
 import numpy as np
-from estimagic.optimization.subsolvers._trsbox_quadratic import minimize_trust_trsbox
 from estimagic.optimization.tranquilo.options import TrustRegion
 from estimagic.optimization.tranquilo.sample_points import _get_effective_bounds
 from estimagic.optimization.tranquilo.sample_points import (
@@ -108,252 +107,3 @@ def log_d_quality_calculator(sample, trustregion, bounds):
     det = np.linalg.det(xtx / n_samples)
     out = n_params * np.log(n_samples) + np.log(det)
     return out
-
-
-# =====================================================================================
-
-
-def get_lambda_poisedness_constant(sample, lower_bounds, upper_bounds):
-    """Calculate the lambda poisedness constant of the sample.
-
-    Lambda-poisedness is a concept to measure how well a set of points is dispersed
-    through a region of interest; here the trust-region. Put differently, the
-    poisedness of a sample reflects how well the sample “spans” the region of
-    interest. Ultimately, it is a measure for how well a model will estimate the
-    criterion function in that region.
-
-    In general, if the sample is lambda-poised with a small lambda, the sample is
-    said to habe "good" geometry. As lambda grows, the system represented by these
-    "points", i.e. vectors, becomes increasingly linearly dependent.
-
-    The metric used for quantifying how well points are positioned in the region
-    of interest is based on Lagrange polynomials.
-
-    Formal definition:
-    A sample Y is said to be lambda-poised on a region of interest if Y is linearly
-    independent and the Lagrange polynomials L(i) of points (vectors) in the
-    sample satisfy:
-
-        lambda >= max_i max_x | L(i) |      (1)
-
-    i.e. for each point i in the sample, we maximize the absolute criterion value
-    of its lagrange polynomial L(i); we then take the maximum over all these
-    criterion values as the lambda constant for this particular sample on the
-    given trust-region.
-
-    Say we compare different samples on the same trust-region, we are usually
-    interested in keeping the sample with the least corresponding value of lambda
-    so that (1) holds. Lambda is >= 1.
-
-    For more details, see Conn et al. (:cite:`Conn2010`).
-
-    Note that the trust-region centered around the origin, and its corresponding
-    radius is normalized to 1.
-
-    Args:
-        sample (np.ndarray): The data sample, shape = (n, p).
-        bounds (Bounds): The parameter bounds.
-
-    Returns:
-        float: The lambda poisedness constant.
-
-    """
-    n_samples, n_params = sample.shape
-
-    interpolation_mat = build_interpolation_matrix(sample.T, n_params, n_samples)
-    lambda_max = -999
-
-    for index in range(n_samples):
-        linear_terms, square_terms = get_lagrange_polynomial(
-            sample=sample,
-            interpolation_mat=interpolation_mat,
-            index=index,
-        )
-
-        x_argmax = minimize_trust_trsbox(
-            linear_terms,
-            square_terms,
-            trustregion_radius=1,
-            lower_bounds=lower_bounds,
-            upper_bounds=upper_bounds,
-        )
-        critval_min = abs(_evaluate_scalar_model(x_argmax, linear_terms, square_terms))
-
-        x_argmin = minimize_trust_trsbox(
-            -linear_terms,
-            -square_terms,
-            trustregion_radius=1,
-            lower_bounds=lower_bounds,
-            upper_bounds=upper_bounds,
-        )
-        critval_max = abs(_evaluate_scalar_model(x_argmin, linear_terms, square_terms))
-
-        if abs(critval_min) >= abs(critval_max):
-            current_critval = critval_min
-        else:
-            current_critval = critval_max
-
-        if current_critval > lambda_max:
-            lambda_max = current_critval
-
-    return lambda_max
-
-
-def maximize_absolute_value_trust_trsbox(
-    lagrange_polynomial,
-    lower_bounds,
-    upper_bounds,
-):
-    """Maximize the absolute value of a Lagrange polynomial in a trust-region setting.
-
-    Let a Lagrange polynomial of degree two be defined by:
-        L(x) = c + g.T @ x + 0.5 x.T @ H @ x,
-
-    where c, g, H denote the intercept, linear terms, and square terms of the
-    scalar model, respectively.
-
-    In order to maximize L(x), we maximize the absolute value of L(x) in a
-    trust-region setting with radius 1. I.e. we solve:
-
-        max_x  abs(c + g.T @ x + 0.5 x.T @ H @ x)
-            s.t. lower_bound <= x <= upper_bound
-                 ||x|| <= 1
-
-    In order to find the solution x*, we both minimize and maximize
-    the objective c + g.T @ x + 0.5 x.T @ H @ x.
-    The resulting candidate vectors are then plugged into the criterion function L(x)
-    to see which one yields the largest absolute criterion value of the Lagrange
-    polynomial.
-
-    Args:
-        lagrange_polynomial (NamedTuple): Named tuple containing the parameters of the
-            lagrange polynomial, i.e.:
-            - ``intercept`` (float): Intercept.
-            - ``linear_terms`` (np.ndarray): 1d array of shape (n,) with linear terms.
-            - ``square_terms`` (np.ndarray): 2d array of shape (n, n) containing
-                quare terms.
-        lower_bounds (np.ndarray): 1d array of shape (n,) with lower bounds
-            for the parameter vector x.
-        upper_bounds (np.ndarray): 1d array of shape (n,) with upper bounds
-            for the parameter vector x.
-
-    Returns:
-        float: The largest absolute criterion value.
-
-    """
-    x_min = minimize_trust_trsbox(
-        lagrange_polynomial.linear_terms,
-        lagrange_polynomial.square_terms,
-        trustregion_radius=1,
-        lower_bounds=lower_bounds,
-        upper_bounds=upper_bounds,
-    )
-    x_max = minimize_trust_trsbox(
-        -lagrange_polynomial.linear_terms,
-        -lagrange_polynomial.square_terms,
-        trustregion_radius=1,
-        lower_bounds=lower_bounds,
-        upper_bounds=upper_bounds,
-    )
-
-    critval_min = abs(_evaluate_scalar_model(x_min, lagrange_polynomial))
-    critval_max = abs(_evaluate_scalar_model(x_max, lagrange_polynomial))
-
-    if abs(critval_min) >= abs(critval_max):
-        critval = critval_min
-    else:
-        critval = critval_max
-
-    return critval
-
-
-def get_lagrange_polynomial(
-    sample,
-    interpolation_mat,
-    index,
-):
-    n_samples, n_params = sample.shape
-
-    rhs = _get_rhs(index, n_samples, n_params)
-    coef, *_ = np.linalg.lstsq(interpolation_mat, rhs, rcond=None)
-
-    if n_samples == n_params:
-        linear_terms = coef
-        square_terms = np.zeros((n_params, n_params))
-    elif n_samples == (n_params + 1) * (n_params + 2) // 2 - 1:
-        linear_terms = coef[:n_params]
-        square_terms = _reshape_coef_to_square_terms(coef[n_params:], n_params)
-    else:
-        linear_terms = coef[n_samples:]
-        square_terms = np.zeros((n_params, n_params))
-        for i in range(n_samples):
-            x = sample[i]
-            square_terms += coef[i] * np.outer(x, x)
-
-    return linear_terms, square_terms
-
-
-def build_interpolation_matrix(sample, n_params, n_samples):
-    if n_samples == n_params:
-        mat = sample.T
-    elif n_samples + 1 == (n_params + 1) * (n_params + 2) // 2:
-        mat = np.empty((n_samples, n_samples))
-        mat[:, :n_params] = sample.T
-        for i in range(n_samples):
-            mat[i, n_params:] = _reshape_mat_to_upper_triangular(
-                np.outer(sample[:, i], sample[:, i])
-                - 0.5 * np.diag(np.square(sample[:, i])),
-                n_params,
-            )
-    else:
-        mat = np.zeros((n_samples + n_params, n_samples + n_params))
-        for i in range(n_samples):
-            for j in range(n_samples):
-                mat[i, j] = 0.5 * np.dot(sample[:, i], sample[:, j]) ** 2
-        mat[:n_samples, n_samples:] = sample.T
-        mat[n_samples:, :n_samples] = sample
-
-    return mat
-
-
-def _get_rhs(index, n_samples, n_params):
-    if n_samples == n_params:
-        rhs = np.zeros(n_params)
-        rhs[index] = 1.0
-    elif n_samples == (n_params + 1) * (n_params + 2) // 2 - 1:
-        rhs = np.zeros(n_samples)
-        rhs[index] = 1.0
-    else:
-        rhs = np.zeros(n_samples + n_params)
-        rhs[index] = 1.0
-
-    return rhs
-
-
-def _reshape_coef_to_square_terms(vec, n_params):
-    mat = np.empty((n_params, n_params))
-    idx = -1
-
-    for j in range(n_params):
-        for i in range(j + 1):
-            idx += 1
-            mat[i, j] = vec[idx]
-            mat[j, i] = vec[idx]
-
-    return mat
-
-
-def _reshape_mat_to_upper_triangular(mat, n_params):
-    triu = np.empty(n_params * (n_params + 1) // 2)
-    idx = -1
-
-    for j in range(n_params):
-        for i in range(j + 1):
-            idx += 1
-            triu[idx] = mat[i, j]
-
-    return triu
-
-
-def _evaluate_scalar_model(x, linear_terms, square_terms):
-    return linear_terms.T @ x + 0.5 * x.T @ square_terms @ x
